@@ -1,12 +1,13 @@
 name: color_picker_example
 description: Simple example of how to use the FlexColorPicker package.
-version: 1.1.3
+version: 2.0.0-nullsafety.0
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0-0 <3.0.0'
 
 dependencies:
-  flex_color_picker: any
+  flex_color_picker:
+    path: ../
 
   flutter:
     sdk: flutter
