@@ -1,10 +1,10 @@
 name: flex_color_picker
 description: A customizable Flutter primary, accent and custom color picker. Includes an optional HSV wheel color picker.
-version: 1.1.3
+version: 2.0.0-nullsafety.0
 homepage: https://github.com/rydmike/flex_color_picker
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0-0 <3.0.0'
 
 dependencies:
   flutter:
@@ -18,5 +18,5 @@ dev_dependencies:
 
   test:
 
-#flutter:
-  #uses-material-design: true
+flutter:
+  uses-material-design: true
